refactor zarrUtils into zarrextra package
