+++ packages/zarrextra/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@spatialdata/core",
+  "name": "@spatialdata/zarrextra",
   "version": "0.0.1",
-  "description": "Core library for interfacing with SpatialData stores",
+  "description": "Extra utilities for working with zarr stores using zarrita",
   "type": "module",
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
@@ -15,30 +15,18 @@
     "dist"
   ],
   "scripts": {
-    "build": "vite build && tsc --emitDeclarationOnly --outDir dist",
+    "build": "vite build",
     "dev": "vite build --watch",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
-    "@spatialdata/zarrextra": "workspace:*",
-    "@math.gl/core": "catalog:",
-    "@zarrita/storage": "^0.1.3",
-    "anndata.js": "catalog:",
-    "apache-arrow": "catalog:",
-    "ol": "^10.6.1",
-    "parquet-wasm": "catalog:",
-    "zarrita": "catalog:",
-    "zod": "catalog:"
-  },
-  "overrides": {
     "zarrita": "catalog:"
   },
   "devDependencies": {
     "@types/node": "^22.10.5",
     "typescript": "catalog:",
-    "unzipit": "^1.4.3",
     "vite": "catalog:",
     "vite-plugin-dts": "catalog:",
     "vitest": "catalog:"
@@ -47,6 +35,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/Taylor-CCB-Group/SpatialData.js.git",
-    "directory": "packages/core"
+    "directory": "packages/zarrextra"
   }
 }
+
